[pad] Split PadOp into dedicated module.
base.py is growing too large, and PadOp is not a small piece of code.
Moreover it embodies a set of classes which may make sense to regroup together down the road.
Best therefore to move it to a separate module already.
