+++ src/pycsou/operator/linop/__init__.py
@@ -1,5 +1,6 @@
 from pycsou.operator.linop.base import *
 from pycsou.operator.linop.kron import *
 from pycsou.operator.linop.nufft import *
+from pycsou.operator.linop.pad import *
 from pycsou.operator.linop.reduce import *
 from pycsou.operator.linop.select import *
