+++ src/pycsou_tests/operator/linop/test_padop.py
